==> R/00_utility_functions.R <==
library(tidyverse)

# read and combine the csvs from all different years
# base_path = path to folder where extracted CSVs are ! Change this path to where
path_to_csv <- file.path("..", "Data", "extracted_data")

read_all_years_csv <- function(base_path=path_to_csv){
  
  df_2019 <- read.csv(paste(base_path, "/2019.csv", sep="")) %>%
    mutate(year = "2019")
  df_2020 <- read.csv(paste(base_path, "/2020.csv", sep="")) %>%
    mutate(year = "2020")
  df_2021 <- read.csv(paste(base_path, "/2021.csv", sep="")) %>%
    mutate(year = "2021")
  df_2022 <- read.csv(paste(base_path, "/2022.csv", sep="")) %>% 
    mutate(year = "2022")
  df_2023 <- read.csv(paste(base_path, "/2023.csv", sep="")) %>%
    mutate(year = "2023")
  
  df_full <- rbind(df_2019, 
                   df_2020, 
                   df_2021, 
                   df_2022, 
                   df_2023)
  
}

# get thresholds for outliers according to tukey
tukey_outlier_threshold<- function(x, k=1.5, bound=c("upper", "lower")){
  if(bound=="upper"){
    return(quantile(x, 0.75, na.rm=TRUE)+(k*IQR(x, na.rm=TRUE)))
  } else if (bound == "lower"){
    return(quantile(x, 0.25, na.rm=TRUE)-(k*IQR(x, na.rm=TRUE)))
  } else {
    warning(paste("bound", bound, "is not implemented. Supply either 'upper' or 'lower'. "))
  }
}


# vectors of CCS topics
top_topics <- c("general and reference",
                "hardware",
                "computer systems organization",
                "networks",
                "software and its engineering",
                "theory of computing",
                "mathematics of computing",
                "information systems",
                "security and privacy",
                "human-centered computing",
                "computing methodologies",
                "applied computing",
                "social and professional topics")

applied_computing_topics <- c(
  "digital cash",
  "e-commerce infrastructure",
  "electronic funds transfer",
  "online shopping",
  "online banking",
  "secure online transactions",
  "online auctions",
  "electronic commerce",
  "intranets",
  "extranets",
  "enterprise resource planning",
  "enterprise applications",
  "data centers",
  "enterprise information systems",
  "business process modeling",
  "business process management systems",
  "business process monitoring",
  "cross-organizational business processes",
  "business intelligence",
  "business process management",
  "enterprise architecture management",
  "enterprise architecture framework",
  "enterprise architecture modeling",
  "enterprise architectures",
  "service-oriented architectures",
  "event-driven architectures",
  "business rules",
  "enterprise modeling",
  "enterprise ontologies, taxonomies and vocabularies",
  "enterprise data management",
  "reference models",
  "business-it alignment",
  "it architectures",
  "it governance",
  "enterprise computing infrastructures",
  "enterprise application integration",
  "information integration and interoperability",
  "enterprise interoperability",
  "enterprise computing",
  "avionics",
  "aerospace",
  "archaeology",
  "astronomy",
  "chemistry",
  "environmental sciences",
  "earth and atmospheric sciences",
  "computer-aided design",
  "engineering",
  "pyhsics",
  "mathematics and statistics",
  "avionics",
  "electronics",
  "internet telephony",
  "Physical sciences and engineering",
  "telecommunications",
  "molecular sequence analysis",
  "recognition of genes and regulatory elements",
  "computational transcriptomics",
  "biological networks",
  "sequencing and genotyping technologies",
  "imaging",
  "computational proteomics",
  "molecular structural biology",
  "computational biology",
  "computational genomics",
  "genomics",
  "systems biology",
  "consumer health",
  "health care information systems",
  "health informatics",
  "bioinformatics",
  "metabolomics / metabonomics",
  "population genetics",
  "computational proteomics",
  "proteomics",
  "transcriptomics",
  "genetics",
  "life and medical sciences",
  "ethnography",
  "anthropogy",
  "law",
  "psychology",
  "sociology",
  "law, social and behavioral sciences",
  "surveillance mechanisms",
  "investigation techniques",
  "evidence collection, storage and analysis",
  "network forensics",
  "system forensics",
  "data recovery",
  "computer forensics",
  "fine arts",
  "performing arts",
  "computer-aided design",
  "architecture (buildings)",
  "language translation",
  "media arts",
  "sound and music computing",
  "arts and humanities",
  "digital libraries and archives",
  "publishing",
  "cyberwarfare",
  "military",
  "cartography",
  "agriculture",
  "voting / election technologies",
  "e-government",
  "computing in government",
  "word processors",
  "spreadsheets",
  "computer games",
  "microcomputers",
  "personal computers and pc applications",
  "computers in other domains",
  "consumer products",
  "supply-chain management",
  "command and control",
  "industry and manufacturing",
  "computer-aided manufacturing",
  "multi-criterion optimization and decision-making",
  "decision analysis",
  "transportation",
  "forecasting",
  "marketing",
  "Operations research",
  "digital libraries and archives",
  "computer-assisted instruction",
  "interactive learning environments",
  "collaborative learning",
  "learning management systems",
  "distance learning",
  "e-learning",
  "computer-managed instruction",
  "Education",
  "document searching",
  "text editing",
  "version control",
  "document metadata",
  "document management",
  "document analysis",
  "document scanning",
  "graphics recognition and interpretation",
  "optical character recognition",
  "online handwriting recognition",
  "document capture",
  "extensible markup language (xml)",
  "markup languages",
  "annotation",
  "format and notation",
  "multi / mixed media creation",
  "image composition",
  "hypertext / hypermedia creation",
  "document scripting languages",
  "document preparation",
  "document management and text processing")

gen_ref_topics <- c("surveys and overviews",
                    "reference works",
                    "biographs",
                    "general literature",
                    "computing standards, rfcs and guidelines",
                    "reliability",
                    "empirical studies",
                    "measurement",
                    "metrics",
                    "evaluation",
                    "experimentation",
                    "estimation",
                    "design",
                    "performance",
                    "validation",
                    "verification",
                    "document types",
                    "cross-computing tools and techniques")


human_centered_comp_topics <- c(
  "user models",
  "user studies",
  "usability testing",
  "heuristic evaluations",
  "laboratory experiments",
  "field studies",
  "hci design and evaluation methods",
  "hypertext / hypermedia",
  "mixed / augmented reality",
  "command line interfaces",
  "graphical user interfaces",
  "virtual reality",
  "web-based interaction",
  "natural language interfaces",
  "collaborative interaction",
  "interaction paradigms",
  "graphics input devices",
  "displays and imagers",
  "sound-based input / output",
  "keyboards",
  "pointing devices",
  "touch screens",
  "haptic devices",
  "interaction devices",
  "hci theory, concepts and models",
  "auditory feedback",
  "text input",
  "pointing",
  "gestural input",
  "interaction techniques",
  "user interface management systems",
  "user interface programming",
  "user interface toolkits",
  "interactive systems and tools",
  "empirical studies in hci",
  "human computer interaction (hci)",
  "user interface design",
  "user centered design",
  "activity centered design",
  "scenario-based design",
  "participatory design",
  "contextual design",
  "interface design prototyping",
  "interaction design process and methods",
  "interaction design theory, concepts and paradigms",
  "empirical studies in interaction design",
  "wireframes",
  "systems and tools for interaction design",
  "interaction design",
  "social content sharing",
  "collaborative content creation",
  "collaborative filtering",
  "social recommendation",
  "social networks",
  "social tagging",
  "computer supported cooperative work",
  "social engineering (social devices)",
  "social navigation",
  "social media",
  "collaborative and social computing theory, concepts and paradigms",
  "social network analysis",
  "ethnographic studies",
  "collaborative and social computing design and evaluation methods",
  "blogs",
  "wikis",
  "reputation systems",
  "open source software",
  "social networking sites",
  "social tagging systems",
  "synchronous editor",
  "asynchronous editors",
  "collaborative and social computing systems and tools",
  "empirical studies in collaborative and social computing",
  "collaborative and social computing devices",
  "collaborative and social computing",
  "ubiquitous computing",
  "mobile computing",
  "ambient intelligence",
  "ubiquitous and mobile computing theory, concepts and paradigms",
  "ubiquitous and mobile computing systems and tools",
  "smartphones",
  "interactive whiteboards",
  "mobile phones",
  "mobile devices",
  "portable media players",
  "personal digital assistants",
  "handheld game consoles",
  "e-book readers",
  "tablet computers",
  "ubiquitous and mobile devices",
  "ubiquitous and mobile computing design and evaluation methods",
  "empirical studies in ubiquitous and mobile computing",
  "ubiquitous and mobile computing",
  "treemaps",
  "hyperbolic trees",
  "heat maps",
  "graph drawings",
  "dendrograms",
  "cladograms",
  "visualization techniques",
  "scientific visualization",
  "visual analytics",
  "geographic visualization",
  "information visualization",
  "visualization application domains",
  "visualization toolkits",
  "visualization systems and tools",
  "visualization theory, concepts and paradigms",
  "empirical studies in visualization",
  "visualization design and evaluation methods",
  "visualization",
  "accessibility theory, concepts and paradigms",
  "empirical studies in accessibility",
  "accessibility design and evaluation methods",
  "accessibility technologies",
  "accessibility systems and tools",
  "accessibility")

security_privacy_topics <- c(
  "key management",
  "digital signatures",
  "public key encryption",
  "public key (asymmetric) techniques",
  "block and stream ciphers",
  "hash functions and message authentication codes",
  "symmetric cryptography and hash functions",
  "cryptoanalysis and other attacks",
  "information-theoretic techniques",
  "mathematical foundations of cryptography",
  "cryotography",
  "trust frameworks",
  "security requirements",
  "formal security models",
  "logic and verification",
  "formal methods and theory of security",
  "biometrics",
  "graphical / visual passwords",
  "multi-factor authentication",
  "authentication",
  "access control",
  "pseudonymity, anonymity and untraceability",
  "privacy-preserving protocols",
  "digital rights management",
  "authorization",
  "security services",
  "malware and its mitigation",
  "artificial immune systems",
  "intrusion detection systems",
  "spoofing attacks",
  "phishing",
  "social engineering attacks",
  "intrusion/anomaly detection and malware mitigation",
  "tamper-proof and tamper-resistant designs",
  "embedded systems security",
  "hardware-based security protocols",
  "hardware security implementation",
  "malicious design modifications",
  "side-channel analysis and countermeasures",
  "hardware attacks and countermeasures",
  "hardware reverse engineering",
  "security in hardware",
  "mobile platform security",
  "trusted computing",
  "virtualization and security",
  "operating systems security",
  "browser security",
  "distributed systems security",
  "information flow control",
  "denial-of-service attacks",
  "firewalls",
  "penetration testing",
  "vulnerability scanners",
  "vulnerability management",
  "file system security",
  "systems security",
  "security protocols",
  "web protocol security",
  "mobile and wireless security",
  "denial-of-service attacks",
  "firewalls",
  "network security",
  "data anonymization and sanitization",
  "management and querying of encrypted data",
  "information accountability and usage control",
  "database activity monitoring",
  "database and storage security",
  "software security engineering",
  "web application security",
  "social network security and privacy",
  "domain-specific security and privacy architectures",
  "software reverse engineering",
  "software and application security",
  "economics of security and privacy",
  "social aspects of security and privacy",
  "privacy protections",
  "usability in security and privacy",
  "human and societal aspects of security and privacy")
